gltf_json::extensions
Module texture

Contains Texture, Sampler, and other related data structures.

Structs

Info

Reference to a Texture.

Sampler

Texture sampler properties for filtering and wrapping modes.

Texture

A texture and its sampler.

TextureTransform

Many techniques can be used to optimize resource usage for a 3d scene. Chief among them is the ability to minimize the number of textures the GPU must load. To achieve this, many engines encourage packing many objects’ low-resolution textures into a single large texture atlas. The region of the resulting atlas that corresponds with each object is then defined by vertical and horizontal offsets, and the width and height of the region.

TextureTransformOffset

The offset of the UV coordinate origin as a factor of the texture dimensions.

TextureTransformRotation

Rotate the UVs by this many radians counter-clockwise around the origin. This is equivalent to a similar rotation of the image clockwise.

TextureTransformScale

The scale factor applied to the components of the UV coordinates.