Struct gltf_json::path::Path

pub struct Path(pub String);

An immutable JSON source path.

Methods

impl Path

pub fn new() -> Self

Creates an empty JSON source path.

Examples

Basic usage

let path = Path::new();
assert_eq!("", path.as_str());

pub fn field(&self, name: &str) -> Self

Returns a new path ending with the given field.

Examples

Basic usage

let path = Path::new().field("foo");
assert_eq!("foo", path.as_str());
assert_eq!("foo.bar", path.field("bar").as_str());

pub fn index(&self, index: usize) -> Self

Returns a new path ending with the given array index.

Examples

Basic usage

let path = Path::new().field("foo");
assert_eq!("foo[123]", path.index(123).as_str());

pub fn key(&self, key: &str) -> Self

Returns a new path ending with the given object key.

Examples

Basic usage

let path = Path::new().field("foo");
assert_eq!("foo[\"bar\"]", path.key("bar").as_str());

pub fn as_str(&self) -> &str

Returns a view into the internal representation.