gltf_json
Module scene

Contains Scene, Node, and other related data structures.

Structs

Node

A node in the node hierarchy. When the node contains skin, all mesh.primitives must contain JOINTS_0 and WEIGHTS_0 attributes. A node can have either a matrix or any combination of translation/rotation/scale (TRS) properties. TRS properties are converted to matrices and postmultiplied in the T * R * S order to compose the transformation matrix; first the scale is applied to the vertices, then the rotation, and then the translation. If none are provided, the transform is the identity. When a node is targeted for animation (referenced by an animation.channel.target), only TRS properties may be present; matrix will not be present.

Scene

The root Nodes of a scene.

UnitQuaternion

Unit quaternion rotation in the order (x, y, z, w), where w is the scalar.