gltf_json
Module serialize

Re-exports of serde_json serialization functions.

This module re-exports the generic serde serialization functions so that one can serialize data structures other than Root without being bound to a specific version of serde_json.

Functions

to_string

Serialize the given data structure as a String of JSON.

to_string_pretty

Serialize the given data structure as a pretty-printed String of JSON.

to_value

Convert a T into serde_json::Value which is an enum that can represent any valid JSON data.

to_vec

Serialize the given data structure as a JSON byte vector.

to_vec_pretty

Serialize the given data structure as a pretty-printed JSON byte vector.

to_writer

Serialize the given data structure as JSON into the I/O stream.

to_writer_pretty

Serialize the given data structure as pretty-printed JSON into the I/O stream.