Module gltf_json::extensions::material

Contains Material and other related data structures.

Structs

AttenuationColor

A colour in the inclusive range [[0.0; 3], [1.0; 3]] with a default value of [1.0; 3].

AttenuationDistance

A number in the inclusive range [0.0, +inf] with a default value of +inf.

EmissiveStrength

EmissiveStrengthFactor

A positive number with 1.0 as the default value.

IndexOfRefraction

A positive number with default value of 1.5

Ior

Material

The material appearance of a primitive.

NormalTexture

Defines the normal texture of a material.

OcclusionTexture

Defines the occlusion texture of a material.

PbrMetallicRoughness

A set of parameter values that are used to define the metallic-roughness material model from Physically-Based Rendering (PBR) methodology.

ThicknessFactor

A number in the inclusive range [0.0, +inf] with a default value of 0.0.

Transmission

TransmissionFactor

A number in the inclusive range [0.0, 1.0] with a default value of 0.0.

Unlit

Empty struct that should be present for primitives which should not be shaded with the PBR shading model.

Volume