Trait gltf_json::validation::Validate

pub trait Validate {
    // Provided method
    fn validate<P, R>(&self, _root: &Root, _path: P, _report: &mut R)
       where P: Fn() -> Path,
             R: FnMut(&dyn Fn() -> Path, Error) { ... }
}

Trait for validating glTF JSON data so that the library can function without panicking.

Provided Methods

fn validate<P, R>(&self, _root: &Root, _path: P, _report: &mut R)
where P: Fn() -> Path, R: FnMut(&dyn Fn() -> Path, Error),

Validates the invariants required for the library to function safely.

Object Safety

This trait is not object safe.