Trait gltf_json::validation::Validate

pub trait Validate {
    fn validate<P, R>(&self, _root: &Root, _path: P, _report: &mut R)
    where
        P: Fn() -> Path,
        R: FnMut(&dyn Fn() -> Path, Error),
    { ... }
}

Trait for validating glTF JSON data so that the library can function without panicking.

Provided methods

fn validate<P, R>(&self, _root: &Root, _path: P, _report: &mut R) where
P: Fn() -> Path,
R: FnMut(&dyn Fn() -> Path, Error),

Validates the invariants required for the library to function safely.

Loading content...